sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xagentauth"
version = "0.1.0"
description = "AgentAuth client SDK for Python — authenticate AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["ai", "authentication", "agent", "jwt"]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2",
    "PyJWT>=2.9",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.3"]
crewai = ["crewai-tools>=0.14"]
fastapi = ["fastapi>=0.100"]
flask = ["flask>=3"]
server = ["fastapi>=0.100", "flask>=3"]
all = ["langchain-core>=0.3", "crewai-tools>=0.14", "fastapi>=0.100", "flask>=3"]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "ruff>=0.8",
    "fastapi>=0.100",
    "flask>=3",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/xagentauth"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 120
